Unite policies POM in single file
