+++ tests/e2e/20-kubewarden.spec.ts
@@ -1,8 +1,6 @@
 import { test, expect } from './rancher-test';
 import { PolicyServersPage } from './pages/policyservers.page';
-import { AdmissionPoliciesPage } from './pages/admissionpolicies.page';
-import { ClusterAdmissionPoliciesPage } from './pages/clusteradmissionpolicies.page';
-import { Policy } from './pages/basepolicypage';
+import { Policy, AdmissionPoliciesPage, ClusterAdmissionPoliciesPage } from './pages/policies.page';
 import { KubewardenPage } from './pages/kubewarden.page';
 
 const expect2m = expect.configure({timeout: 2*60_000})
