+++ tests/e2e/pages/policies.page.ts
@@ -125,3 +125,17 @@ export abstract class BasePolicyPage extends BasePage {
   }
 
 }
+
+export class AdmissionPoliciesPage extends BasePolicyPage {
+  async goto(): Promise<void> {
+    // await this.nav.explorer('Kubewarden', 'AdmissionPolicies')
+    await this.page.goto('dashboard/c/local/kubewarden/policies.kubewarden.io.admissionpolicy')
+  }
+}
+
+export class ClusterAdmissionPoliciesPage extends BasePolicyPage {
+  async goto(): Promise<void> {
+    // await this.nav.explorer('Kubewarden', 'ClusterAdmissionPolicies')
+    await this.page.goto('dashboard/c/local/kubewarden/policies.kubewarden.io.clusteradmissionpolicy')
+  }
+}
